Fixe issue : get the rights course ids of membership order.

+++ admin/class-lifterlms-discord-addon-admin.php
@@ -693,24 +693,34 @@ class Lifterlms_Discord_Addon_Admin {
 	}
 
 	/**
+	 * Assign Or Remove discord roles on update an Order.
 	 *
+	 * @param     int     $order_id Order ID.
+	 * @param WP_POST $obj Post Object.
+	 * @param bool    $update  Whether this is an existing post being updated.
 	 */
 	public function ets_lifterlms_change_order_status( $order_id, $obj, $update ) {
 
 		$order        = new LLMS_Order( $order_id );
 		$order_status = $order->get_access_status();
+
 		if ( $order_status == 'inactive' ) {
 			$user_id                            = $order->get( 'user_id' );
-			$course_id                          = $order->get( 'product_id' );
+			$membership = new LLMS_Membership( $order->get( 'product_id' ) );
+			$course_ids = $membership->get_auto_enroll_courses();
 			$ets_lifterlms_discord_role_mapping = json_decode( get_option( 'ets_lifterlms_discord_role_mapping' ), true );
 			$access_token                       = sanitize_text_field( trim( get_user_meta( $user_id, '_ets_lifterlms_discord_access_token', true ) ) );
 			$refresh_token                      = sanitize_text_field( trim( get_user_meta( $user_id, '_ets_lifterlms_discord_refresh_token', true ) ) );
 			if ( $access_token && $refresh_token ) {
-				if ( is_array( $ets_lifterlms_discord_role_mapping ) && array_key_exists( 'course_id_' . $course_id, $ets_lifterlms_discord_role_mapping ) ) {
-					$discord_role = sanitize_text_field( trim( $ets_lifterlms_discord_role_mapping[ 'course_id_' . $course_id ] ) );
-					if ( $discord_role && $discord_role != 'none' ) {
-						delete_user_meta( $user_id, '_ets_lifterlms_discord_role_id_for_' . $course_id, $discord_role );
-						$this->lifterlms_discord_public_instance->delete_discord_role( $user_id, $discord_role );
+				if ( is_array( $course_ids ) && count( $course_ids ) ) {
+					foreach ( $course_ids as $course_id ){
+						if ( is_array( $ets_lifterlms_discord_role_mapping ) && array_key_exists( 'course_id_' . $course_id, $ets_lifterlms_discord_role_mapping ) ) {
+							$discord_role = sanitize_text_field( trim( $ets_lifterlms_discord_role_mapping[ 'course_id_' . $course_id ] ) );
+							if ( $discord_role && $discord_role != 'none' ) {
+								delete_user_meta( $user_id, '_ets_lifterlms_discord_role_id_for_' . $course_id, $discord_role );
+								$this->lifterlms_discord_public_instance->delete_discord_role( $user_id, $discord_role );
+							}
+						}
 					}
 				}
 			}
